homeworks/2025/06-streaming/pyproject.toml
[project]
name = "06-streaming"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-flink>=1.20.1",
    "kafka-python>=2.1.1",
    "pgcli>=4.2.0",
    "polars>=1.25.2",
]
